% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AmunSentinelProcessing.R
\name{Contrast}
\alias{Contrast}
\title{Function to adjust the contrast of an image}
\usage{
Contrast(band, min, max)
}
\arguments{
\item{band}{Band to be adjusted}

\item{min}{new min value}

\item{max}{new max value}
}
\value{
new scaled image
}
\description{
Function to adjust the contrast of an image
}
