% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AmunSentinelProcessing.R
\name{ContrastRGB}
\alias{ContrastRGB}
\title{Plots RGB with new contrast}
\usage{
ContrastRGB(red, green, blue, rmin, rmax, gmin, gmax, bmin, bmax)
}
\arguments{
\item{red}{redBand}

\item{green}{greenBand}

\item{blue}{blueBand}

\item{rmin}{RedBand min value}

\item{rmax}{redBand max value}

\item{gmin}{greenBand min value}

\item{gmax}{greenBand max value}

\item{bmin}{blueBand min value}

\item{bmax}{blueBand max value}
}
\value{
returns RGB image with new Contrast
}
\description{
Plots RGB with new contrast
}
