% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AmunSentinelProcessing.R
\name{getContrastSeperatly}
\alias{getContrastSeperatly}
\title{Display the image with new contrast}
\usage{
getContrastSeperatly(band, min, max)
}
\arguments{
\item{band}{Band to be displayed}

\item{min}{min contrast value}

\item{max}{max contrast value}
}
\value{
Displays the image with new Contrast
}
\description{
Display the image with new contrast
}
