% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AmunSentinelProcessing.R
\name{NDVI}
\alias{NDVI}
\title{returns values based on NDVI formula (NIR-red)/(NIR+red)}
\usage{
NDVI(x, y)
}
\arguments{
\item{x}{NIR band}

\item{y}{Red band}
}
\description{
returns values based on NDVI formula (NIR-red)/(NIR+red)
}
