% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AmunSentinelProcessing.R
\name{NDVI_Result}
\alias{NDVI_Result}
\title{Returns the NDVI image}
\usage{
NDVI_Result(x, y)
}
\arguments{
\item{x}{NIR Band}

\item{y}{Red band}
}
\value{
return NDVI image
}
\description{
Returns the NDVI image
}
