% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AmunSentinelProcessing.R
\name{TCI}
\alias{TCI}
\title{TCI Function}
\usage{
TCI(channel1, channel2, channel3)
}
\arguments{
\item{channel1}{First Channel of the image}

\item{channel2}{Second Channel of the image}

\item{channel3}{Third Channel of the image}
}
\value{
returns an RGB image consisting of the input channels as R,G and B
}
\description{
TCI Function
}
