% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AmunSentinelProcessing.R
\name{TCI_grayscale}
\alias{TCI_grayscale}
\title{Delivers the TCI image as grayscale}
\usage{
TCI_grayscale(img)
}
\arguments{
\item{img}{The image which will be displayed as grayscale}
}
\value{
returns grayscale image
}
\description{
Delivers the TCI image as grayscale
}
